Conversational_AI/requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-classic
python-dotenv
duckduckgo-search
ddgs
